+++ NEWS.md
@@ -2,6 +2,7 @@ Known issues: https://github.com/PredictiveEcology/NetLogoR/issues
 
 version 0.3.0
 =============
+* `inRadius` now multiplies the `width` by a tiny amount so that the function returns an inclusive result.
 * add `quickPlot >= 0.1.1.9000` dependency
 * update dependencies on `SpaDES`-related packages to only include the ones actually used.
 * Define [[ and $ for `worldArray` to extract subset layers.